Move mieq composite index tests (#5729)
